Bugfix: registered keys not set bug
This is a bugfix proposal for the authority keys not being generated and stored.

The function verify_keys call from the OauthManager->storeKey function has been
commented. Passing the third option to the verify_keys function issued
the exception and blocked the registration process. I am not sure this
is the best fix. The function verify_keys should be patched but I am not
sure how the main developer intended it to work.

The OauthManager->storeKey function has been patched as well.
It was refering to the unknown attribute keys.

Also, the checking for the already stored keys was broken.
This patch proposes a solution by first checking for the existing
database entry.